+++ packages/zeebe-element-templates-json-schema/test/fixtures/job-priority-definition/valid-message-end-event.js
@@ -1,0 +1,25 @@
+export const template = {
+  name: 'Job Priority Definition',
+  id: 'job-priority-definition-message-end',
+  description: 'A template to define job priority on a message end event.',
+  version: 1,
+  appliesTo: [
+    'bpmn:EndEvent'
+  ],
+  elementType: {
+    value: 'bpmn:EndEvent',
+    eventDefinition: 'bpmn:MessageEventDefinition'
+  },
+  properties: [
+    {
+      value: '50',
+      type: 'Hidden',
+      binding: {
+        type: 'zeebe:jobPriorityDefinition',
+        property: 'priority'
+      }
+    }
+  ]
+};
+
+export const errors = null;