+++ packages/zeebe-element-templates-json-schema/CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to [@camunda/zeebe-element-templates-json-schema](https://gi
 
 ___Note:__ Yet to be released changes appear here._
 
+## 0.43.0
+
+* `FEAT`: add `zeebe:jobPriorityDefinition` binding type
+
 ## 0.42.0
 
 * `FEAT`: support `isEmpty` condition ([#252](https://github.com/camunda/element-templates-json-schema/pull/252))